requirements.txt
deta==1.1.0
geopy==2.2.0
meteostat==1.6.5
numpy==1.23.3
pandas==1.5.0
plotly==5.10.0
pygithub==1.55
python-dotenv==0.21.0
scipy==1.9.1
streamlit-authenticator==0.2.1
streamlit==1.13.0
streamlit-lottie==0.0.3